restructured the directory structure for better resusability
