+++ www/main/common/store/user-settings/user-settings.json
@@ -1,0 +1,244 @@
+{
+  "categories": {
+    "slide-layout": {
+      "title": "SLIDE_LAYOUT",
+      "type": "title",
+      "subcategories": ["font-sizes", "display-options", "autoplay-options"]
+    },
+    "display-options": {
+      "title": "DISPLAY_OPTIONS",
+      "type": "switch",
+      "settings": [
+        "akhandpatt",
+        "display-next-line",
+        "left-align",
+        "larivaar",
+        "larivaar-assist",
+        "larivaar-assist-type",
+        "display-vishraams",
+        "vishraam-type",
+        "vishraam-source"
+      ]
+    },
+    "font-sizes": {
+      "title": "FONT_SIZES",
+      "type": "range",
+      "max": 20,
+      "min": 1,
+      "step": 1,
+      "settings": [
+        "gurbani-font-size",
+        "translation-font-size",
+        "transliteration-font-size",
+        "teeka-font-size",
+        "announcements-font-size"
+      ]
+    },
+    "visibility": {
+      "title": "VISIBILITY",
+      "type": "checkbox",
+      "settings": ["translation-visibility", "transliteration-visibility", "teeka-visibility"]
+    },
+    "autoplay-options": {
+      "title": "AUTOPLAY_OPTIONS",
+      "type": "switch",
+      "settings": ["autoplay-toggle", "autoplay-delay"]
+    },
+    "bani-and-languages": {
+      "title": "BANI_AND_LANGUAGES",
+      "type": "title",
+      "subcategories": ["bani", "language"]
+    },
+    "bani": {
+      "title": "BANI_SETTINGS",
+      "type": "dropdown",
+      "settings": ["bani-length", "mangal-position"]
+    },
+    "language": {
+      "title": "LANGUAGE_SETTINGS",
+      "type": "dropdown",
+      "settings": ["translation-language", "transliteration-language"]
+    },
+    "app-settings": {
+      "title": "APP_SETTINGS",
+      "type": "title",
+      "subcategories": ["app-layout", "other-options"]
+    },
+    "app-layout": {
+      "title": "APP_LAYOUT",
+      "type": "switch",
+      "settings": ["presenter-view", "quick-tools", "shortcut-tray", "live-feed"]
+    },
+    "other-options": {
+      "title": "OTHER_OPTIONS",
+      "type": "switch",
+      "settings": ["limit-change-log", "statistics"]
+    }
+  },
+  "settings": {
+    "gurbani-font-size": {
+      "title": "BANI",
+      "initialValue": 9
+    },
+    "translation-font-size": {
+      "title": "TRANSLATION",
+      "initialValue": 5
+    },
+    "teeka-font-size": {
+      "title": "PUNJABI_TEEKA",
+      "initialValue": 4
+    },
+    "transliteration-font-size": {
+      "title": "TRANSLITERATION",
+      "initialValue": 4
+    },
+    "announcements-font-size": {
+      "title": "ANNOUNCEMENTS",
+      "initialValue": 7
+    },
+    "translation-visibility": {
+      "title": "TRANSLATION_VISIBILITY",
+      "type": "checkbox",
+      "initialValue": true
+    },
+    "teeka-visibility": {
+      "title": "TEEKA_VISIBILITY",
+      "type": "checkbox",
+      "initialValue": true
+    },
+    "transliteration-visibility": {
+      "title": "TRANSLITERATION_VISIBILITY",
+      "type": "checkbox",
+      "initialValue": true
+    },
+    "akhandpatt": {
+      "title": "AKHAND_PATH_VIEW",
+      "notes": "CHROMECAST_UNAVAILABLE",
+      "initialValue": false
+    },
+    "display-next-line": {
+      "title": "NEXT_LINE",
+      "initialValue": false
+    },
+    "left-align": {
+      "title": "LEFT_ALIGN",
+      "initialValue": false
+    },
+    "larivaar": {
+      "title": "LARIVAAR",
+      "initialValue": false
+    },
+    "larivaar-assist": {
+      "title": "LARIVAAR_ASSIST",
+      "notes": "LARIVAAR_ASSIST_MSG",
+      "initialValue": false
+    },
+    "larivaar-assist-type": {
+      "title": "ASSIST_TYPE",
+      "type": "dropdown",
+      "options": {
+        "single-color": "SINGLE_COLOR",
+        "multi-color": "MULTI_COLOR"
+      },
+      "initialValue": "single-color"
+    },
+    "display-vishraams": {
+      "title": "SHOW_VISHRAAMS",
+      "initialValue": false
+    },
+    "vishraam-type": {
+      "title": "VISHRAAM_OPTIONS",
+      "type": "dropdown",
+      "initialValue": "colored-words",
+      "options": {
+        "colored-words": "COLORED_WORDS",
+        "gradient-bg": "GRADIENT_BG"
+      }
+    },
+    "vishraam-source": {
+      "title": "VISHRAAM_SOURCE",
+      "type": "dropdown",
+      "options": {
+        "sttm2": "STTM2",
+        "igurbani": "I_GURBANI",
+        "sttm": "STTM"
+      },
+      "initialValue": "sttm-2"
+    },
+    "autoplay-toggle": {
+      "title": "AUTO_PLAY",
+      "initialValue": false
+    },
+    "autoplay-delay": {
+      "title": "DELAY",
+      "type": "range",
+      "max": 20,
+      "min": 1,
+      "step": 1,
+      "initialValue": 10
+    },
+    "bani-length": {
+      "title": "BANI_LENGTH",
+      "type": "dropdown",
+      "options": {
+        "short": "SHORT_BANI_LENGTH",
+        "medium": "MEDIUM_BANI_LENGTH",
+        "long": "LONG_BANI_LENGTH",
+        "extralong": "EXTRA_LONG_BANI_LENGTH"
+      },
+      "initialValue": "short"
+    },
+    "mangal-position": {
+      "title": "MANGAL_POSITION",
+      "type": "dropdown",
+      "options": {
+        "current": "CURRENT_MANGAL",
+        "above": "ABOVE_MANGAL"
+      },
+      "initialValue": "current"
+    },
+    "translation-language": {
+      "title": "TRANSLATION_LANGUAGE",
+      "type": "dropdown",
+      "options": {
+        "English": "ENGLISH_LANGUAGE",
+        "Spanish": "SPANISH_LANGUAGE"
+      },
+      "initialValue": "English"
+    },
+    "transliteration-language": {
+      "title": "TRANSLITERATION_LANGUAGE",
+      "type": "dropdown",
+      "options": {
+        "English": "ENGLISH_LANGUAGE",
+        "Devanagri": "DEVANAGARI_LANGUAGE"
+      },
+      "initialValue": "English"
+    },
+    "presenter-view": {
+      "title": "PRESENTER_VIEW",
+      "initialValue": false
+    },
+    "quick-tools": {
+      "title": "QUICK_TOOLS",
+      "initialValue": false
+    },
+    "shortcut-tray": {
+      "title": "SHORTCUT_TRAY",
+      "initialValue": false
+    },
+    "live-feed": {
+      "title": "LIVE_FEED",
+      "initialValue": false
+    },
+    "limit-change-log": {
+      "title": "LIMIT_CHANGELOG",
+      "initialValue": false,
+      "notes": "LIMIT_CHANGELOG_MSG"
+    },
+    "statistics": {
+      "title": "COLLECT_STATISTICS",
+      "initialValue": false
+    }
+  }
+}
