--- www/main/common/utils/save-settings.js
@@ -1,11 +1,0 @@
-/*const saveSettings = (settingKey, payload) => {
-  savedSettings[settingKey] = payload;
-  fs.writeFileSync(userConfigPath, JSON.stringify(savedSettings));
-
-  // Update localStorage for viewer
-  if (typeof localStorage === 'object') {
-    localStorage.setItem('prefs', JSON.stringify(savedSettings));
-  }
-
-  console.log(savedSettings);
-}*/
